Create user support

+++ habit/ResetProgress.aspx.cs
@@ -12,35 +12,39 @@ namespace habit
     {
         protected void Page_Load(object sender, EventArgs e)
         {
-            System.IO.DirectoryInfo di = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory + "files\\Habits");
-
-            foreach (FileInfo file in di.GetFiles())
+            if (Session["LoginUser"] != null)
             {
-                if (file.Extension == ".xml")
+                if (Session["LoginUser"].ToString() != "")
                 {
-                    if (file.Name.StartsWith(Session["LoginUser"] + "_"))
+                    System.IO.DirectoryInfo di = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory + "files\\Habits");
+
+                    foreach (FileInfo file in di.GetFiles())
                     {
-                        file.Delete();
+                        if (file.Extension == ".xml")
+                        {
+                            if (file.Name.StartsWith(Session["LoginUser"] + "_"))
+                            {
+                                file.Delete();
+                            }
+                        }
+
+                    }
+
+                    System.IO.DirectoryInfo di2 = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory + "files\\Users\\User");
+
+                    foreach (FileInfo file in di2.GetFiles())
+                    {
+                        if (file.Extension == ".xml")
+                        {
+                            if (file.Name == Session["LoginUser"] + ".xml")
+                            {
+                                file.Delete();
+                            }
+                        }
+
                     }
                 }
-
             }
-
-            System.IO.DirectoryInfo di2 = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory + "files\\Users\\User");
-
-            foreach (FileInfo file in di2.GetFiles())
-            {
-                if (file.Extension == ".xml")
-                {
-                    if (file.Name == Session["LoginUser"] + ".xml")
-                    {
-                        file.Delete();
-                    }
-                }
-
-            }
-
-
             Response.Redirect("/");
         }
     }
